+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ next
 * Union members that come after `str` or `pathlib.Path` no longer need to have
   a valid parser.
 * Parsed function docstring made available as ``signature(...).doc``.
+* Removed the deprecated ``strict_kwonly`` in favor of ``cli_options``.
 
 6.4.0 (2022-07-19)
 ------------------